third_party/blink/web_tests/wpt_internal/speech/scripted/speechrecognition-errors.html

<!doctype html>
<title>Test Speech JavaScript API errors</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="/gen/third_party/blink/public/mojom/speech/speech_recognizer.mojom.js"></script>
<script src="../resources/mock-speechrecognizer.js"></script>
<script>

async_test(t => {
  var r = new webkitSpeechRecognition();

  for (var prop in r) {
    if (prop.match('^on')) {
      r[prop] = function() {
        assert_unreached('unexpected ' + event.type + ' event!');
      }
    }
  }
  var count = 0;
  r.start();
  mockSpeechRecognizer.setMockSpeechRecognitionError(
    blink.mojom.SpeechRecognitionErrorCode.kNotAllowed);

  // Check that we get an error event.
  r.onerror = t.step_func(event => {
    assert_equals(count, 0);
    count++;
    assert_equals(event.error, 'not-allowed');
    assert_equals(event.message, '');
    assert_equals(event.type, 'error');
  });

  // Check that we get an end event after the error event.
  r.onend = t.step_func_done(() => {
    assert_equals(count, 1);
  });
}, 'Not-allowed error test');

</script>
